Added crops to Crop.js and added User.js and index.js to /models
