+++ models/index.js
@@ -1,0 +1,4 @@
+const User = require('./User');
+const Crop = require('./Crop');
+
+module.exports = { User, Crop };
